@@ -3,20 +3,24 @@
 
 'use strict';
 
-import { inject, injectable } from 'inversify';
+import { inject, injectable, named } from 'inversify';
 import { DiagnosticSeverity } from 'vscode';
 import { STANDARD_OUTPUT_CHANNEL } from '../../common/constants';
 import { ILogger, IOutputChannel } from '../../common/types';
+import { displayProgress } from '../../common/utils/decorators';
+import { Diagnostics } from '../../common/utils/localize';
 import { IServiceContainer } from '../../ioc/types';
 import { IApplicationDiagnostics } from '../types';
 import { IDiagnostic, IDiagnosticsService, ISourceMapSupportService } from './types';
 
 @injectable()
 export class ApplicationDiagnostics implements IApplicationDiagnostics {
-    constructor(@inject(IServiceContainer) private readonly serviceContainer: IServiceContainer) { }
+    constructor(@inject(IServiceContainer) private readonly serviceContainer: IServiceContainer,
+        @inject(IOutputChannel) @named(STANDARD_OUTPUT_CHANNEL) private readonly outputChannel: IOutputChannel) { }
     public register() {
         this.serviceContainer.get<ISourceMapSupportService>(ISourceMapSupportService).register();
     }
+    @displayProgress(Diagnostics.progress())
     public async performPreStartupHealthCheck(): Promise<void> {
         const diagnosticsServices = this.serviceContainer.getAll<IDiagnosticsService>(IDiagnosticsService);
         await Promise.all(diagnosticsServices.map(async diagnosticsService => {
@@ -29,18 +33,17 @@ export class ApplicationDiagnostics implements IApplicationDiagnostics {
     }
     private log(diagnostics: IDiagnostic[]): void {
         const logger = this.serviceContainer.get<ILogger>(ILogger);
-        const outputChannel = this.serviceContainer.get<IOutputChannel>(IOutputChannel, STANDARD_OUTPUT_CHANNEL);
         diagnostics.forEach(item => {
             const message = `Diagnostic Code: ${item.code}, Message: ${item.message}`;
             switch (item.severity) {
                 case DiagnosticSeverity.Error: {
                     logger.logError(message);
-                    outputChannel.appendLine(message);
+                    this.outputChannel.appendLine(message);
                     break;
                 }
                 case DiagnosticSeverity.Warning: {
                     logger.logWarning(message);
-                    outputChannel.appendLine(message);
+                    this.outputChannel.appendLine(message);
                     break;
                 }
                 default: {
