@@ -1,4 +1,5 @@
-import { isTestExecution, isUnitTestExecution } from '../constants';
+import { ProgressLocation, ProgressOptions, window } from 'vscode';
+import { isTestExecution } from '../constants';
 // tslint:disable-next-line:no-require-imports no-var-requires
 const _debounce = require('lodash/debounce') as typeof import('lodash/debounce');
 
@@ -13,7 +14,7 @@ export function debounce(wait?: number) {
     return function (_target: any, _propertyName: string, descriptor: TypedPropertyDescriptor<any>) {
         const originalMethod = descriptor.value!;
         // If running tests, lets not debounce (so tests run fast).
-        wait = wait && isUnitTestExecution() ? undefined : wait;
+        wait = wait && isTestExecution() ? undefined : wait;
         // tslint:disable-next-line:no-invalid-this no-any
         (descriptor as any).value = _debounce(function () { return originalMethod.apply(this, arguments); }, wait);
     };
@@ -51,6 +52,25 @@ export function swallowExceptions(scopeName: string) {
                 }
                 console.error(errorMessage, error);
             }
+        };
+    };
+}
+
+// tslint:disable-next-line:no-any
+type PromiseFunction = (...any: any[]) => Promise<any>;
+
+export function displayProgress(title: string, location = ProgressLocation.Window) {
+    return function (_target: Object, _propertyName: string, descriptor: TypedPropertyDescriptor<PromiseFunction>) {
+        const originalMethod = descriptor.value!;
+        // tslint:disable-next-line:no-any no-function-expression
+        descriptor.value = async function (...args: any[]) {
+            const progressOptions: ProgressOptions = { location, title };
+            // tslint:disable-next-line:no-invalid-this
+            const promise = originalMethod.apply(this, args);
+            if (!isTestExecution()) {
+                window.withProgress(progressOptions, () => promise);
+            }
+            return promise;
         };
     };
 }
