@@ -9,6 +9,7 @@ import { EXTENSION_ROOT_DIR } from '../../constants';
 
 // External callers of localize use these tables to retrieve localized values.
 export namespace Diagnostics {
+    export const progress = localize('diagnostics.progress', 'Validating user setup');
     export const warnSourceMaps = localize('diagnostics.warnSourceMaps', 'Source map support is enabled in the Python Extension, this will adversely impact performance of the extension.');
     export const disableSourceMaps = localize('diagnostics.disableSourceMaps', 'Disable Source Map Support');
     export const warnBeforeEnablingSourceMaps = localize('diagnostics.warnBeforeEnablingSourceMaps', 'Enabling source map support in the Python Extension will adversely impact performance of the extension.');
@@ -19,10 +20,11 @@ export namespace Common {
     export const canceled = localize('Common.canceled', 'Canceled');
 }
 
-export namespace LanguageServiceSurveyBanner {
-    export const bannerMessage = localize('LanguageServiceSurveyBanner.bannerMessage', 'Can you please take 2 minutes to tell us how the Python Language Server is working for you?');
-    export const bannerLabelYes = localize('LanguageServiceSurveyBanner.bannerLabelYes', 'Yes, take survey now');
-    export const bannerLabelNo = localize('LanguageServiceSurveyBanner.bannerLabelNo', 'No, thanks');
+export namespace LanguageService {
+    export const bannerMessage = localize('LanguageService.bannerMessage', 'Can you please take 2 minutes to tell us how the Python Language Server is working for you?');
+    export const bannerLabelYes = localize('LanguageService.bannerLabelYes', 'Yes, take survey now');
+    export const bannerLabelNo = localize('LanguageService.bannerLabelNo', 'No, thanks');
+    export const progressActivating = localize('LanguageService.progressActivating', 'Activating Language Server');
 }
 
 export namespace Interpreters {
