@@ -13,8 +13,8 @@ import * as localize from '../../client/common/utils/localize';
 suite('localize tests', () => {
 
     test('keys', done => {
-        const val = localize.LanguageServiceSurveyBanner.bannerMessage();
-        assert.equal(val, 'Can you please take 2 minutes to tell us how the Python Language Server is working for you?', 'LanguageServiceSurveyBanner string doesnt match');
+        const val = localize.LanguageService.bannerMessage();
+        assert.equal(val, 'Can you please take 2 minutes to tell us how the Python Language Server is working for you?', 'LanguageService string doesnt match');
         done();
     });
 
@@ -22,7 +22,7 @@ suite('localize tests', () => {
         // Force a config change
         process.env.VSCODE_NLS_CONFIG = '{ "locale": "it" }';
 
-        const val = localize.LanguageServiceSurveyBanner.bannerLabelYes();
+        const val = localize.LanguageService.bannerLabelYes();
         assert.equal(val, 'Sì, prenderò il sondaggio ora', 'bannerLabelYes is not being translated');
         done();
     });
